requirements.txt
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
opencv-python==4.9.0.80
flask==3.0.2
matplotlib==3.8.3
pandas==2.2.1
pillow==10.2.0
scikit-learn==1.4.1.post1
jupyter==1.0.0
